Module sudoku_backend::ops

Modules

constraints
Various verbose weak type-system-level constraints.

errors
Various API errors.

routes

setup
Data used for general application setup: third-party connexions, et caetera.

tables
diesel table definitions, copied over from doc/umentation.

Structs

BoardMessage
The message sent to and from the client on acquisiion/submission of solved board

LoginForm
Log-in and user-create form – refer to doc/user.md.

NewBoardRequestForm
Form requested by the client to play a new Sudoku board

OldBoardRequestForm
Form requested by the client to play the specified Sudoku board

SanitisedUserData
Data returned to authenticating client, with a free conversion from a User.

Session
Refer to doc/session.md for more details.

SudokuBoard
Refer to doc/sudoku.md for more details.

SudokuSolution
Refer to doc/sudoku.md for more details.

User
Refer to doc/user.md for more details.

Enums

BoardDifficulty
The difficulties sudoku boards can come in.

LeaderboardOf
Which leaderboard to acquire.

SolutionOrdering
How to order a solution list.