Module sudoku_backend::ops::constraints

Various verbose weak type-system-level constraints.

Structs

EMail
Weakly validate the string to be roughly e-mail-shaped.

HexString
Require that the string be hexadecimal and of the specified width.

NonEmpty
Require that the string not be empty.

Positive
Weakly validate the int to be positive.

SCrypt64Length
Character length of a SCrypt key with dkLen=64,

SudokuBoard9x9ConciseLength
Character length of the concise representation of a 9x9 Sudoku board.

SudokuString
Require that the string be a concise sudoku representation of the specified width.

Traits

StringLength
Template value arg, refer to HexString.