Struct sudoku_backend::ops::Session

pub struct Session {
    pub id: Option<i32>,
    pub expiry: NaiveDateTime,
    pub is_admin: bool,
    pub user_id: Option<i32>,
}

Refer to doc/session.md for more details.

TODO: doc full workflow with creation, database access, etc.

Fields
id: Option<i32>

Unique user ID.

Actually not optional, but this allows us to get an ID from the database.

expiry: NaiveDateTime

Time the cookie expires, defaults to a day from creation time.

is_admin: bool

Whether the user is logged in as admin (always false if user_id == None).

user_id: Option<i32>

The logged-in user ID, or None for unauthed access.

Methods

impl Session

pub fn new() -> Session

Create an empty session expiring a day from the creation datetime.

Put it in the database to get an ID to put back in the cookie.

pub fn log_in(
&mut self,
uid: i32,
is_admin: bool,
db: &SqliteConnection
) -> Result<(), &'static str>

Update the in-memory and in-DB model to be logged in a the specified user with the specified permissions.

Examples

Before:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 0, NULL);

Update:

let mut sess = Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
};

sess.log_in(18, true, &db).unwrap();

assert_eq!(sess, Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: true,
    user_id: Some(18),
});

After:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 1, 18);

pub fn log_out(&mut self, db: &SqliteConnection) -> Result<(), &'static str>

Update the in-memory and in-DB model to not be logged in nor be admin.

Examples

Before:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 1, 18);

Update:

let mut sess = Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: true,
    user_id: Some(18),
};

sess.log_out(&db).unwrap();

assert_eq!(sess, Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
});

After:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 0, NULL);