Struct sudoku_backend::ops::User

pub struct User {
    pub id: Option<i32>,
    pub username: String,
    pub password: String,
    pub email: String,
    pub created_at: NaiveDateTime,
    pub is_admin: bool,
    pub points_total: i32,
}

Refer to doc/user.md for more details.

Fields
id: Option<i32>

Unique user ID.

Actually not optional, but this allows us to get an ID from the database.

username: String

User's name or "login" or whatever.

password: String

Doubly scrypted password, see doc/user.md for details.

email: String

User's contact e-mail

created_at: NaiveDateTime

Time user was created.

is_admin: bool

Whether the user has administrative privileges.

points_total: i32

Sum total of the user's points, calculated according to doc/scoring.md#endgame-formula, CHECKed to nonnegativity.

Methods

impl User

pub fn new<U: Into<String>, P: Into<String>, E: Into<String>>(
username: U,
password: P,
email: E
) -> User

Create a defaulted user with the specified parameters.

pub fn get_by_id(id: i32, db: &SqliteConnection) -> Result<User, &'static str>

Pull out a user with the specified ID from the specified database.

pub fn add(&self, db: &SqliteConnection) -> Result<(), &'static str>

Insert this user into the specified database, if possible.