Struct sudoku_backend::ops::setup::DatabaseConnection

pub struct DatabaseConnection(_);

Connection request guard type: a wrapper around an r2d2 pooled connection.

Use this as an argument to a Rocket request handler after having it manage()d to gain access to the database.

Examples

#[get("/databased")]
fn databased(db: DatabaseConnection) -> String {
    // Do some database ops, which are outside the scope of this document
    funxion_result
}

fn main() {
    let database_file: (String, PathBuf) = /* obtained elsewhere */;
    rocket::ignite()
        .manage(DatabaseConnection::initialise(&database_file))
        .mount("/", routes![databased])
        .launch();
}

Methods

impl DatabaseConnection

pub fn initialise(
db_file: &(String, PathBuf)
) -> Pool<ConnectionManager<SqliteConnection>>

Set up a connection to the main database located in the specified file and initialise it with util::INITIALISE_DATABASE.