Struct sudoku_backend::ops::Session

pub struct Session {
    pub id: Option<i32>,
    pub expiry: NaiveDateTime,
    pub is_admin: bool,
    pub user_id: Option<i32>,
    pub sudoku_board_id: Option<i32>,
    pub board_skeleton: Option<Cow<'static, str>>,
    pub solve_start: Option<NaiveDateTime>,
}

Refer to doc/session.md for more details.

Fields
id: Option<i32>

Unique user ID.

Actually not optional, but this allows us to get an ID from the database.

expiry: NaiveDateTime

Time the cookie expires, defaults to a day from creation time.

is_admin: bool

Whether the user is logged in as admin (always false if user_id == None).

user_id: Option<i32>

The logged-in user ID, or None for unauthed access.

sudoku_board_id: Option<i32>

ID of board currently being solved

board_skeleton: Option<Cow<'static, str>>

The board skeleton sent to the user

solve_start: Option<NaiveDateTime>

Time the solving started

Methods

impl Session

pub fn new() -> Session

Create an empty session expiring a day from the creation datetime.

Put it in the database to get an ID to put back in the cookie.

pub fn get(
db: &SqliteConnection,
cookies: &mut Cookies
) -> Result<Session, &'static str>

Get/create a session for the specified cookieset.

pub fn start_game<S: Into<Cow<'static, str>>>(
&mut self,
bid: i32,
skeleton: S,
db: &SqliteConnection
) -> Result<(), &'static str>

Update the in-memory and in-DB model to start playing the specified game.

Examples

let mut sess = Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
    sudoku_board_id: None,
    board_skeleton: None,
    solve_start: None,
};

let skeleton = "3....5....25.81.........6......4..73.3....14..1.7......8...9.12.97........2..8.54";
sess.start_game(24, skeleton, &db).unwrap();

assert_eq!(sess, Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
    sudoku_board_id: Some(24),
    board_skeleton: Some(skeleton.into()),
    solve_start: Some(NaiveDateTime::from_timestamp(Utc::now().timestamp(), 0)),
});

pub fn stop_game(&mut self, db: &SqliteConnection) -> Result<(), &'static str>

Update the in-memory and in-DB model to start playing the specified game.

Examples

let mut sess = Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
    sudoku_board_id: Some(24),
    board_skeleton: Some("3....5....25.81.........6......4..73.3....14..1.7......8...9.12.97........2..8.54".into()),
    solve_start: Some(NaiveDate::from_ymd(2018, 7, 9).and_hms(2, 41, 27)),
};

sess.stop_game(&db).unwrap();

assert_eq!(sess, Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
    sudoku_board_id: None,
    board_skeleton: None,
    solve_start: None,
});

pub fn game_started(&self) -> bool

Check if this session has a game going

pub fn log_in(
&mut self,
uid: i32,
is_admin: bool,
db: &SqliteConnection
) -> Result<(), &'static str>

Update the in-memory and in-DB model to be logged in as the specified user with the specified permissions.

Examples

Before:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 0, NULL, NULL, NULL, NULL);

Update:

let mut sess = Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
    sudoku_board_id: None,
    board_skeleton: None,
    solve_start: None,
};

sess.log_in(18, true, &db).unwrap();

assert_eq!(sess, Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: true,
    user_id: Some(18),
    sudoku_board_id: None,
    board_skeleton: None,
    solve_start: None,
});

After:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 1, 18, NULL, NULL, NULL);

pub fn log_out(&mut self, db: &SqliteConnection) -> Result<(), &'static str>

Update the in-memory and in-DB model to not be logged in nor be admin.

Examples

Before:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 1, 18, NULL, NULL, NULL);

Update:

let mut sess = Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: true,
    user_id: Some(18),
    sudoku_board_id: None,
    board_skeleton: None,
    solve_start: None,
};

sess.log_out(&db).unwrap();

assert_eq!(sess, Session {
    id: Some(32),
    expiry: NaiveDate::from_ymd(2018, 7, 9).and_hms(12, 40, 26),
    is_admin: false,
    user_id: None,
    sudoku_board_id: None,
    board_skeleton: None,
    solve_start: None,
});

After:

INSERT INTO "sessions" VALUES(32, '2018-07-09 12:40:26', 0, NULL, NULL, NULL, NULL);