Struct sudoku_backend::ops::User

pub struct User {
    pub id: Option<i32>,
    pub username: String,
    pub password: String,
    pub email: String,
    pub created_at: NaiveDateTime,
    pub is_admin: bool,
    pub points_total: i32,
    pub games_total: i32,
    pub games_total_easy: i32,
    pub games_total_medium: i32,
    pub games_total_hard: i32,
}

Refer to doc/user.md for more details.

Fields
id: Option<i32>

Unique user ID.

Actually not optional, but this allows us to get an ID from the database.

username: String

User's name or "login" or whatever.

password: String

Doubly scrypted password, see doc/user.md for details.

email: String

User's contact e-mail

created_at: NaiveDateTime

Time user was created.

is_admin: bool

Whether the user has administrative privileges.

points_total: i32

Sum total of the user's points, calculated according to doc/scoring.md#endgame-formula, CHECKed to nonnegativity.

games_total: i32

Total amount of games played, CHECKed to nonnegativity.

games_total_easy: i32

Amount of easy games played, CHECKed to nonnegativity.

games_total_medium: i32

Amount of medium games played, CHECKed to nonnegativity.

games_total_hard: i32

Amount of hard games played, CHECKed to nonnegativity.

Methods

impl User

pub fn new<U: Into<String>, P: Into<String>, E: Into<String>>(
username: U,
password: P,
email: E
) -> User

Create a defaulted user with the specified parameters.

pub fn get_by_id(id: i32, db: &SqliteConnection) -> Result<User, &'static str>

Pull out a user with the specified ID from the specified database.

pub fn add(&self, db: &SqliteConnection) -> Result<(), &'static str>

Insert this user into the specified database, if possible.

pub fn solve(
&mut self,
for_points: usize,
for_difficulty: BoardDifficulty,
db: &SqliteConnection
) -> Result<(), &'static str>

Update in-memory and in-DB repr by the specified point count.

pub fn leaders(
cfg: &LeaderboardConfig,
db: &SqliteConnection
) -> Result<Vec<User>, &'static str>

Retrieve the board with the specified ID.

Examples

Given:

INSERT INTO "users"
    VALUES(1, 'karolsw3',       'password', 'email', '2018-07-23 18:18:24', 0, 435, 1, 1, 0, 0);
INSERT INTO "users"
    VALUES(2, 'nabijaczleweli', 'password', 'email', '2018-07-23 19:08:09', 1, 732, 1, 0, 1, 0);
INSERT INTO "users"
    VALUES(3, 'sehe',           'password', 'email', '2018-07-23 19:08:56', 0, 1230, 2, 0, 0, 2);
INSERT INTO "users"
    VALUES(4, 'skorezore',      'password', 'email', '2018-07-23 19:11:06', 0, 222, 1, 0, 0, 1);

The following holds:

let users = User::leaders(&LeaderboardConfig {
    count: 3,
    ordering: SolutionOrdering::WorstToBest,
}, &db).unwrap();
assert_eq!(
    users,
    &[User {
          id: Some(4),
          username: "skorezore".to_string(),
          password: "password".to_string(),
          email: "skorezore@e.mail".to_string(),
          created_at: NaiveDate::from_ymd(2018, 07, 23).and_hms(19, 11, 06),
          is_admin: false,
          points_total: 222,
          games_total: 1,
          games_total_easy: 0,
          games_total_medium: 0,
          games_total_hard: 1,
      },
      User {
          id: Some(1),
          username: "karolsw3".to_string(),
          password: "password".to_string(),
          email: "karolsw3@e.mail".to_string(),
          created_at: NaiveDate::from_ymd(2018, 07, 23).and_hms(18, 18, 24),
          is_admin: false,
          points_total: 435,
          games_total: 1,
          games_total_easy: 1,
          games_total_medium: 0,
          games_total_hard: 0,
      },
      User {
          id: Some(2),
          username: "nabijaczleweli".to_string(),
          password: "password".to_string(),
          email: "nabijaczleweli@e.mail".to_string(),
          created_at: NaiveDate::from_ymd(2018, 07, 23).and_hms(19, 08, 09),
          is_admin: true,
          points_total: 732,
          games_total: 1,
          games_total_easy: 0,
          games_total_medium: 1,
          games_total_hard: 0,
      }]);